GCP: improve integration docs
Update README with new sections about how to configure authentication,
ingration settings, and how to start collecting logs creating a pub/sub topic,
a subscription and a sink.

The fields and eventa sample are included in both the README and the
dedicated .md files, in preparation of making them visible on the main
docs site https://docs.elastic.co/en/integrations/gcp.